Config-breaking update (see notes)
This update adds Spotify authorization to increase the rate limit at which the bot can request Spotify information. Since the bot will now be looking for an extra set of credentials, the config file needs to be regenerated. The recommended way of doing this is like so:

1. Rename `credentials.ini` to something else, like `credentials.ini.old`
2. Open this renamed file in your text editor of choice.
3. Start the bot. It will fail to find the configuration file and begin the generation sequence.
4. Copy and paste the values from the old config file into the console where appropriate.
5. Enter the new Spotify credentials when prompted.
6. Once the new `credentials.ini` file has been written, delete the old one.

Alternatively, you can open up your existing `credentials.ini` and add the following section to it, filling in information where appropriate:

[S]
c = [CLIENT-ID HERE]
s = [SECRET HERE]